Update X-Stream-Client header to add the device and OS version on requests.
